+++ Models/PDL.cs
@@ -6,25 +6,25 @@ namespace DataSorting.Models
 	public abstract class APDL
 	{
 		public abstract int n { get; set; }
-		public abstract float[] arr { get; set; }
+		public abstract double[] arr { get; set; }
 		public abstract int randMax { get; set; }
 		public abstract Random rnd { get; set; }
-		public Dictionary<string, float> parameters { get; set; }
-		public abstract float GenValue();
+		public Dictionary<string, double> parameters { get; set; }
+		public abstract double GenValue();
 		public abstract void FillArr();
 	}
 
 	public class Fisk : APDL
 	{
 		public override int n { get; set; }
-		public override float[] arr { get; set; }
+		public override double[] arr { get; set; }
 		public override int randMax { get; set; }
 		public override Random rnd { get; set; }
 
-		public float A { get; set; }
+		public double A { get; set; }
 
-		float b;
-		public float B
+		double b;
+		public double B
 		{
 			get { return b; }
 			set
@@ -35,8 +35,8 @@ namespace DataSorting.Models
 			}
 		}
 
-		float c;
-		public float C
+		double c;
+		public double C
 		{
 			get { return c; }
 			set
@@ -50,13 +50,13 @@ namespace DataSorting.Models
 		public Fisk() {
 			randMax = 32767;
 			rnd = new Random();
-			parameters = new Dictionary<string, float>();
+			parameters = new Dictionary<string, double>();
 		}
 
-		public override float GenValue()
+		public override double GenValue()
 		{
-			float randomNumber = (float)rnd.Next(randMax) / randMax;
-			float fiskValue = A + B / (float)Math.Pow(1.0 / randomNumber - 1, 1.0 / C);
+			double randomNumber = (double)rnd.Next(randMax) / randMax;
+			double fiskValue = A + B / (double)Math.Pow(1.0 / randomNumber - 1, 1.0 / C);
 			return fiskValue;
 		}
 
@@ -65,7 +65,7 @@ namespace DataSorting.Models
 			A = parameters[nameof(A)];
 			B = parameters[nameof(B)];
 			C = parameters[nameof(C)];
-			arr = new float[n];
+			arr = new double[n];
 			for (int i = 0; i < n; i++)
 				arr[i] = GenValue();
 		}
